# docs/bike_easyfinder_pj/SystemDesign/mvp-openapi-v1.yaml
openapi: 3.1.0
info:
  title: Bike EasyFinder MVP API
  version: 1.1.0
  description: 九州7県を対象とする限定β版の公開API契約。
servers:
  - url: https://api.example.invalid/v1
tags:
  - name: Recommendations
  - name: Interactions
  - name: Privacy
paths:
  /recommendations:
    post:
      tags: [Recommendations]
      operationId: createRecommendations
      summary: 現在地と利用条件からツーリング候補を返す
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationRequest'
      responses:
        '200':
          description: 候補がない場合も空配列で返す。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
        '400': {$ref: '#/components/responses/BadRequest'}
        '422': {$ref: '#/components/responses/UnprocessableEntity'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
        '504': {$ref: '#/components/responses/GatewayTimeout'}
  /interactions:
    post:
      tags: [Interactions]
      operationId: recordInteractions
      summary: 明示同意済みの匿名行動イベントを受け付ける
      parameters:
        - $ref: '#/components/parameters/InstallationId'
        - $ref: '#/components/parameters/DeletionTokenHash'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionBatch'
      responses:
        '202':
          description: event_id単位で冪等に受理した件数。
          content:
            application/json:
              schema:
                type: object
                required: [accepted]
                properties:
                  accepted: {type: integer, minimum: 0}
        '400': {$ref: '#/components/responses/BadRequest'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /installations/{installation_id}/data:
    delete:
      tags: [Privacy]
      operationId: requestInstallationDataDeletion
      summary: 匿名installationに紐づくデータ削除を受け付ける
      description: 存在やトークン一致を推測できないよう常に同じ202応答を返す。
      parameters:
        - name: installation_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - name: X-Deletion-Token
          in: header
          required: true
          schema: {type: string, minLength: 32}
        - $ref: '#/components/parameters/RequestId'
      responses:
        '202': {description: 削除要求受理。該当データは7日以内に削除する。}
        '429': {$ref: '#/components/responses/TooManyRequests'}
components:
  parameters:
    InstallationId:
      name: X-Installation-ID
      in: header
      required: true
      description: 分析同意後に端末で生成する匿名UUID。
      schema: {type: string, format: uuid}
    DeletionTokenHash:
      name: X-Deletion-Token-Hash
      in: header
      required: true
      description: 端末で生成した削除トークンの小文字SHA-256。生トークンは削除時だけ送信する。
      schema: {type: string, pattern: '^[0-9a-f]{64}$'}
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      schema: {type: string, format: uuid}
  schemas:
    Coordinate:
      type: object
      additionalProperties: false
      required: [latitude, longitude]
      properties:
        latitude: {type: number, minimum: -90, maximum: 90}
        longitude: {type: number, minimum: -180, maximum: 180}
    Interest:
      type: string
      enum: [any, scenic, sea, mountain, winding, cafe, food, onsen, roadside_station, night_view, historic]
    SpotCategory:
      type: string
      enum: [scenic_view, coast, mountain, cafe, restaurant, onsen, roadside_station, historic_site, tourist_attraction]
    SpotReaction:
      type: object
      additionalProperties: false
      required: [spot_id, reaction]
      properties:
        spot_id: {type: string, format: uuid}
        reaction: {type: string, enum: [interested, not_interested, visited]}
    PreferenceProfile:
      type: object
      additionalProperties: false
      properties:
        spot_reactions:
          type: array
          maxItems: 200
          items: {$ref: '#/components/schemas/SpotReaction'}
        category_weights:
          type: object
          additionalProperties: {type: integer, minimum: -30, maximum: 15}
        tag_weights:
          type: object
          additionalProperties: {type: integer, minimum: -30, maximum: 15}
    RecommendationRequest:
      type: object
      additionalProperties: false
      required: [origin, available_minutes, interests, allow_highway]
      properties:
        origin: {$ref: '#/components/schemas/Coordinate'}
        available_minutes: {type: integer, enum: [60, 120, 180, 240, 360, 600]}
        interests:
          type: array
          minItems: 1
          uniqueItems: true
          items: {$ref: '#/components/schemas/Interest'}
        allow_highway: {type: boolean}
        locale: {type: string, default: ja-JP}
        preference_profile: {$ref: '#/components/schemas/PreferenceProfile'}
    RecommendationResponse:
      type: object
      required: [recommendation_id, generated_at, recommendation_rule_version, data_version, candidates, warnings]
      properties:
        recommendation_id: {type: string, format: uuid}
        generated_at: {type: string, format: date-time}
        recommendation_rule_version: {type: string}
        data_version: {type: string}
        candidates:
          type: array
          maxItems: 5
          items: {$ref: '#/components/schemas/RecommendationCandidate'}
        warnings:
          type: array
          items: {$ref: '#/components/schemas/Warning'}
    RecommendationCandidate:
      type: object
      required: [spot_id, name, category, tags, coordinate, outbound_minutes, stay_minutes, return_minutes, return_buffer_minutes, estimated_total_minutes, distance_km, score, reasons, route, freshness, constraints]
      properties:
        spot_id: {type: string, format: uuid}
        name: {type: string}
        summary: {type: [string, 'null']}
        category: {$ref: '#/components/schemas/SpotCategory'}
        tags:
          type: array
          items: {$ref: '#/components/schemas/Interest'}
        coordinate: {$ref: '#/components/schemas/Coordinate'}
        outbound_minutes: {type: integer, minimum: 0}
        stay_minutes: {type: integer, minimum: 0}
        return_minutes: {type: integer, minimum: 0}
        return_buffer_minutes: {type: integer, const: 15}
        estimated_total_minutes: {type: integer, minimum: 0}
        return_eta: {type: [string, 'null'], format: date-time}
        distance_km: {type: number, minimum: 0}
        score: {type: number, minimum: 0, maximum: 100}
        reasons: {type: array, minItems: 1, maxItems: 3, items: {type: string}}
        route: {$ref: '#/components/schemas/RouteSummary'}
        freshness: {$ref: '#/components/schemas/Freshness'}
        constraints: {$ref: '#/components/schemas/Constraints'}
    RouteSummary:
      type: object
      required: [profile, encoded_polyline, shape_format]
      properties:
        profile: {type: string, const: auto}
        encoded_polyline: {type: string}
        shape_format: {type: string, const: polyline6}
    Freshness:
      type: object
      required: [spot_verified_at, route_generated_at]
      properties:
        spot_verified_at: {type: string, format: date-time}
        route_generated_at: {type: string, format: date-time}
        hours_verified_at: {type: [string, 'null'], format: date-time}
    Constraints:
      type: object
      required: [opening_hours, weather, road_information]
      properties:
        opening_hours: {type: string, enum: [open, closed, unknown]}
        weather: {type: string, enum: [not_requested, ok, warning, unavailable]}
        road_information: {type: string, const: external_confirmation_required}
    Warning:
      type: object
      required: [code, message]
      properties:
        code: {type: string, enum: [PARTIAL_ROUTE_FAILURE, STALE_SPOT_DATA, HOURS_UNKNOWN, WEATHER_UNAVAILABLE, ROAD_INFORMATION_EXTERNAL]}
        message: {type: string}
    InteractionBatch:
      type: object
      additionalProperties: false
      required: [events]
      properties:
        events: {type: array, minItems: 1, maxItems: 50, items: {$ref: '#/components/schemas/InteractionEvent'}}
    InteractionEvent:
      type: object
      additionalProperties: false
      required: [event_id, session_id, recommendation_id, spot_id, event_type, occurred_at]
      properties:
        event_id: {type: string, format: uuid}
        session_id: {type: string, format: uuid}
        recommendation_id: {type: string, format: uuid}
        spot_id: {type: string, format: uuid}
        event_type: {type: string, enum: [shown, interested, not_interested, visited, selected, route_started]}
        occurred_at: {type: string, format: date-time}
        rank: {type: [integer, 'null'], minimum: 1, maximum: 5}
        available_minutes: {type: [integer, 'null'], enum: [60, 120, 180, 240, 360, 600, null]}
        interests: {type: array, items: {$ref: '#/components/schemas/Interest'}}
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        code: {type: string}
        detail: {type: string}
        request_id: {type: string, format: uuid}
  responses:
    BadRequest:
      description: JSONまたはパラメータが不正。
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    UnprocessableEntity:
      description: 値は解釈できるが業務制約に違反。
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    PayloadTooLarge:
      description: 1回50件を超過。
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    TooManyRequests:
      description: レート上限超過。
      headers:
        Retry-After: {schema: {type: integer}}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    ServiceUnavailable:
      description: 一時的な依存サービス障害。
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    GatewayTimeout:
      description: 推薦処理が4秒の期限を超過。
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}